Adds support for having multiple fields be returned from an INSERT.
The returning method must now accept either a scalar string or an array
of values.  I had intended more string arguments for this library, but
I know if I didn't allow for both it would be the source of too many
implementation errors.

Added a new unit test for the multiple returning fields.

The point of this is to allow for multiple primary key tables.  Even if
I think they're a bad idea, I still want the support there for them.