Allows scalar values to be bound when calling where()
As I've been using this library, the most common usage is having a
single value being compared to a field when calling where().  This has
also been the most common problem, as I often forget to wrap the binding
value in array brackets.

With this change, an implementer to call with a simple scalar value or
an array of values and the result will work as expected either way.

I tweaked on a couple of existing unit tests to make sure this change
is exercised.