Provide a new Where object with 3 major sub types
So this can all fit into the existing API for how where clauses are now
added, a single object will handle the 3 known types of where clauses
that are supported.  Each type has its own class in order to deal with
their unique requirements.

Most importantly, the Where\Select object can now take in a sub select
object that won't render until the time the SQL is generated.  This
means that bindings will need to be accounted for at the time of the
final rendering as well.

issue #1